+++ rxbinding-design-kotlin/src/main/kotlin/com/jakewharton/rxbinding/support/design/widget/RxAppBarLayout.kt
@@ -10,11 +10,3 @@ import rx.Observable
  * to free this reference.
  */
 public inline fun AppBarLayout.offsetChanges(): Observable<Int> = RxAppBarLayout.offsetChanges(this)
-
-/**
- * Create an observable which emits offsetChange events for the `view`.
- * 
- * *Warning:* The created observable keeps a strong reference to `view`. Unsubscribe
- * to free this reference.
- */
-public inline fun AppBarLayout.offsetChangeEvents(): Observable<AppBarLayoutOffsetChangeEvent> = RxAppBarLayout.offsetChangeEvents(this)