+++ rxbinding-kotlin/src/main/kotlin/com/jakewharton/rxbinding/widget/RxCompoundButton.kt
@@ -18,19 +18,6 @@ import rx.functions.Action1
 public inline fun CompoundButton.checkedChanges(): Observable<Boolean> = RxCompoundButton.checkedChanges(this)
 
 /**
- * Create an observable of checked change events for `view`.
- * 
- * *Warning:* The created observable keeps a strong reference to `view`. Unsubscribe
- * to free this reference.
- * 
- * *Warning:* The created observable uses [CompoundButton.setOnCheckedChangeListener]
- * to observe checked changes. Only one observable can be used for a view at a time.
- * 
- * *Note:* A value will be emitted immediately on subscribe.
- */
-public inline fun CompoundButton.checkedChangeEvents(): Observable<CompoundButtonCheckedChangeEvent> = RxCompoundButton.checkedChangeEvents(this)
-
-/**
  * An action which sets the checked property of `view`.
  * 
  * *Warning:* The created observable keeps a strong reference to `view`. Unsubscribe