+++ rxbinding-kotlin/src/main/kotlin/com/jakewharton/rxbinding/widget/RxRadioGroup.kt
@@ -15,16 +15,6 @@ import rx.functions.Action1
 public inline fun RadioGroup.checkedChanges(): Observable<Int> = RxRadioGroup.checkedChanges(this)
 
 /**
- * Create an observable of the checked view ID change events in `view`.
- * 
- * *Warning:* The created observable keeps a strong reference to `view`. Unsubscribe
- * to free this reference.
- * 
- * *Note:* A value will be emitted immediately on subscribe.
- */
-public inline fun RadioGroup.checkedChangeEvents(): Observable<RadioGroupCheckedChangeEvent> = RxRadioGroup.checkedChangeEvents(this)
-
-/**
  * An action which sets the checked child of `view` with ID. Passing {@code -1} will clear
  * any checked view.
  *